+++ templates/ice-react-material-template/template/scaffold/ice.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  entry: 'src/index.js',
+  publicPath: './',
+  plugins: [
+    ['ice-plugin-fusion', {
+      themePackage: '@icedesign/theme',
+    }],
+    ['ice-plugin-moment-locales', {
+      locales: ['zh-cn'],
+    }],
+  ],
+};