+++ tools/ice-scripts/index.js
@@ -1,9 +1,0 @@
-const dev = require('./lib/dev');
-const build = require('./lib/build');
-const init = require('./lib/init');
-
-module.exports = {
-  dev,
-  build,
-  init,
-};