+++ packages/ice-npm-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 1.1.2
+
+- [fix] 修复getAndExtractTarball写空文件时卡死的问题
+
+## 1.1.1
+
+- [fix] 用 request-promise 替换 axios 之后的参数变化
+
+## 1.1.0
+
+- [feat] 新增 getAndExtractTarball&getNpmTarball 两个 API
+- [chore] 请求库从 axios 统一为 request
+
 ## 1.0.3
 
 - [chore] 增加 log